+++ scandeval/benchmarks/abstract/base.py
@@ -71,6 +71,9 @@ class BaseBenchmark(ABC):
             label, and `id2label[split_point]` contains the labels for the
             second label. Only relevant if `two_labels` is True. Defaults to
             None.
+        prefer_jax (bool, optional):
+            Whether to prefer Jax for the benchmarking, if available. Defaults
+            to False.
         verbose (bool, optional):
             Whether to print additional output during evaluation. Defaults to
             False.
@@ -87,6 +90,7 @@ class BaseBenchmark(ABC):
         cache_dir (str): Directory where models are cached.
         two_labels (bool): Whether two labels should be predicted.
         split_point (int or None): Splitting point of `id2label` into labels.
+        prefer_jax (bool): Whether to prefer Jax for the benchmarking.
         verbose (bool): Whether to print additional output.
     '''
     def __init__(self,
@@ -99,6 +103,7 @@ class BaseBenchmark(ABC):
                  cache_dir: str = '.benchmark_models',
                  two_labels: bool = False,
                  split_point: Optional[int] = None,
+                 prefer_jax: bool = False,
                  verbose: bool = False):
 
         self.short_name = name
@@ -113,6 +118,7 @@ class BaseBenchmark(ABC):
         self.cache_dir = cache_dir
         self.two_labels = two_labels
         self.split_point = split_point
+        self.prefer_jax = prefer_jax
         self.verbose = verbose
 
         if id2label is not None:
@@ -696,9 +702,14 @@ class BaseBenchmark(ABC):
                       for a in a_tags_with_class
                       if 'tag-red' in a['class']]
 
+        # Set up the order of the frameworks
+        if self.prefer_jax:
+            valid_frameworks = ['jax', 'pytorch', 'tensorflow', 'spacy']
+        else:
+            valid_frameworks = ['pytorch', 'tensorflow', 'jax', 'spacy']
+
         # Extract a single valid framework in which the model has been
         # implemented
-        valid_frameworks = ['pytorch', 'tensorflow', 'jax', 'spacy']
         for valid_framework in valid_frameworks:
             if valid_framework in frameworks:
                 framework = valid_framework